+++ google/src/test/scala/org/broadinstitute/dsde/workbench/google/GoogleUtilitiesSpec.scala
@@ -35,6 +35,7 @@ class GoogleUtilitiesSpec extends TestKit(ActorSystem("MySpec")) with GoogleUtil
     val errInfo = new ErrorInfo()
 
     message foreach httpExc.setMessage
+    message foreach errInfo.setMessage
     reason foreach errInfo.setReason
     domain foreach errInfo.setDomain
 
@@ -67,6 +68,7 @@ class GoogleUtilitiesSpec extends TestKit(ActorSystem("MySpec")) with GoogleUtil
   }
 
   "when500orGoogleError" should "return true for 500 or Google errors" in {
+    when500orGoogleError(buildGoogleJsonResponseException(403, Some("Unable to extract resource containers."), None, Some("forbidden"))) shouldBe true
     when500orGoogleError(buildGoogleJsonResponseException(403, None, None, Some("usageLimits"))) shouldBe true
     when500orGoogleError(buildGoogleJsonResponseException(429, None, None, Some("usageLimits"))) shouldBe true
     when500orGoogleError(buildGoogleJsonResponseException(400, None, Some("invalid"), None)) shouldBe true
@@ -86,7 +88,7 @@ class GoogleUtilitiesSpec extends TestKit(ActorSystem("MySpec")) with GoogleUtil
   it should "return false otherwise" in {
     when500orGoogleError(buildGoogleJsonResponseException(400, None, Some("boom"), None)) shouldBe false
     when500orGoogleError(buildGoogleJsonResponseException(401)) shouldBe false
-    when500orGoogleError(buildGoogleJsonResponseException(403, None, None, Some("boom"))) shouldBe false
+    when500orGoogleError(buildGoogleJsonResponseException(403, Some("boom"), None, Some("boom"))) shouldBe false
     when500orGoogleError(buildGoogleJsonResponseException(429, None, None, Some("boom"))) shouldBe false
 
     when500orGoogleError(buildHttpResponseException(400)) shouldBe false