+++ google/CHANGELOG.md
@@ -10,6 +10,7 @@ SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.18
 - Moved `SamModel`
 - add googleFirestore support
 - add DistributedLock backed by google firestore
+- add functions to access `DefaultObjectAccessControls` and `BucketAccesscontrols`
 
 ## 0.17
 - upgrade cats to 1.4.0 and scala to 2.12.7